Moved Conflcit command in JEI module
