+++ CraftTweaker2-MC1120-Mod-JEI/src/main/java/crafttweaker/mods/jei/JEIAddonPlugin.java
@@ -16,6 +16,7 @@ public class JEIAddonPlugin implements IModPlugin {
     public static IIngredientRegistry itemRegistry;
     public static IRecipeRegistry recipeRegistry;
     public static IIngredientHelper<ItemStack> ingredientHelper;
+    public static IJeiRuntime jeiRuntime;
     
     @Override
     public void registerItemSubtypes(ISubtypeRegistry subtypeRegistry) {
@@ -42,6 +43,7 @@ public class JEIAddonPlugin implements IModPlugin {
     @Override
     public void onRuntimeAvailable(IJeiRuntime iJeiRuntime) {
         recipeRegistry = iJeiRuntime.getRecipeRegistry();
+        jeiRuntime = iJeiRuntime;
     }
     
     public static List<ItemStack> getSubTypes(ItemStack stack) {
