+++ CraftTweaker2-MC1120-Mod-JEI/src/main/java/crafttweaker/mods/jei/JEIMod.java
@@ -1,9 +1,15 @@
 package crafttweaker.mods.jei;
 
 import crafttweaker.CraftTweakerAPI;
+import crafttweaker.mc1120.commands.*;
+import crafttweaker.mods.jei.commands.ConflictCommand;
+import net.minecraft.command.ICommandSender;
+import net.minecraft.server.MinecraftServer;
 import net.minecraftforge.fml.common.*;
-import net.minecraftforge.fml.common.event.FMLLoadCompleteEvent;
+import net.minecraftforge.fml.common.event.*;
 
+import static crafttweaker.mc1120.commands.SpecialMessagesChat.getClickableCommandText;
+import static crafttweaker.mc1120.commands.SpecialMessagesChat.getNormalMessage;
 import static crafttweaker.mods.jei.JEI.LATE_HIDING;
 
 @Mod(modid = "crafttweakerjei", name = "CraftTweaker JEI Support", version = "2.0.0", dependencies = "after:jei;")
@@ -19,5 +25,27 @@ public class JEIMod {
                 e.printStackTrace();
             }
         }
+    }
+    
+    @Mod.EventHandler
+    public void init(FMLInitializationEvent event){
+        
+        //region >>> JEI Command
+        if(Loader.isModLoaded("jei")) {
+            CTChatCommand.registerCommand(new ConflictCommand());
+        } else {
+            CTChatCommand.registerCommand(new CraftTweakerCommand("conflict") {
+                @Override
+                protected void init() {
+                    setDescription(getClickableCommandText("\u00A72/ct conflict", "/ct conflict", true), getNormalMessage(" \u00A73Lists all conflicting crafting recipes in the game"), getNormalMessage(" \u00A73Might take a bit of time depending on the size of the pack"), getNormalMessage(" \u00A73This needs to be run on a client and with JEI installed"));
+                }
+            
+                @Override
+                public void executeCommand(MinecraftServer server, ICommandSender sender, String[] args) {
+                    sender.sendMessage(getNormalMessage("\u00A74This Command needs to be run with JEI installed"));
+                }
+            });
+        }
+        //endregion
     }
 }
